+++ src/tests/fixtures/elementFixture.ts
@@ -15,7 +15,7 @@ const elementBase: Omit<ExcalidrawElement, "type"> = {
   roughness: 1,
   opacity: 100,
   groupIds: [],
-  strokeSharpness: "sharp",
+  roundness: null,
   seed: 1041657908,
   version: 120,
   versionNonce: 1188004276,